install.sh: Improve UX

+++ api/install.sh
@@ -2,6 +2,26 @@ clear
 
 ARROW=$(tput bold)$(tput setaf 6; echo -n '==>'; tput sgr0;)
 
+IS_HOMEBREW_INSTALLED=false
+if hash brew 2>/dev/null; then
+  IS_HOMEBREW_INSTALLED=true
+fi
+
+IS_NVM_INSTALLED=false
+if ! [ -x "$(command -v nvm)" ]; then
+  IS_NVM_INSTALLED=true
+fi
+
+IS_NODE_INSTALLED=false
+if hash node 2>/dev/null; then
+  IS_NODE_INSTALLED=true
+fi
+
+IS_VSCODE_INSTALLED=false
+if hash code 2>/dev/null; then
+  IS_VSCODE_INSTALLED=true
+fi
+
 echo "Welcome to the installer!"
 echo -e "First, introduce your password to execute all the commands as super user. \n"
 echo -e "\033[1;31mImportant:\033[0m You can be asked more times for password during the process."
@@ -10,23 +30,6 @@ echo
 
 sudo -v
 while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
-
-install_homebrew() {
-    if hash brew 2>/dev/null; then
-        echo "Homebrew already installed!"
-    else
-        ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-    fi
-}
-
-update_homebrew() {
-  brew update
-  brew upgrade
-}
-
-install_git() {
-  brew install git
-}
 
 install_bundle() {
   brew tap caskroom/cask
@@ -73,23 +76,41 @@ install_npm_packages() {
   npm install -g "${npm_packages[@]}"
 }
 
-read -ep "${ARROW} Install Homebrew and update it? [y/n]: "
+#----------------------------
+# Homebrew
+#----------------------------
+if $IS_HOMEBREW_INSTALLED; then
+  echo "${ARROW} Homebrew already installed!"
+else
+  read -ep "${ARROW} Install Homebrew and update it? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing Homebrew..."
-  install_homebrew
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing Homebrew..."
+    ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
-  echo "${ARROW} Updating Homebrew formulas..."
-  update_homebrew
+    echo "${ARROW} Updating Homebrew formulas..."
+    brew update
+    brew upgrade
+
+    IS_HOMEBREW_INSTALLED=true
+  fi
 fi
 
-read -p "${ARROW} Install latest Git version? [y/n]: "
+#----------------------------
+# Git
+#----------------------------
+if $IS_HOMEBREW_INSTALLED; then
+  read -p "${ARROW} Install latest Git version via Homebrew? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing Git..."
-  install_git
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing Git..."
+    brew install git
+  fi
 fi
 
+#----------------------------
+# .bash_profile
+#----------------------------
 read -p "${ARROW} Configure .bash_profile? [y/n]: "
 
 if [ "$REPLY" == "y" ]; then
@@ -97,16 +118,25 @@ if [ "$REPLY" == "y" ]; then
   cp .bash_profile ~ 
 fi
 
-read -p "${ARROW} Install bundle of applications? [y/n]: "
+#----------------------------
+# Application bundle
+#----------------------------
+if $IS_HOMEBREW_INSTALLED; then
+  read -p "${ARROW} Install bundle of applications via Homebrew-Cask? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing applications..."
-  install_bundle
-  echo "${ARROW} Installing applications..."
-  open /Applications/Flux.app
-  open /Applications/Spectacle.app
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing applications..."
+    install_bundle
+
+    open /Applications/Flux.app
+    open /Applications/Spectacle.app
+    open /Applications/KeepingYouAwake.app
+  fi
 fi
 
+#----------------------------
+# Terminal profile
+#----------------------------
 read -p "${ARROW} Configure Terminal profile? [y/n]: "
 
 if [ "$REPLY" == "y" ]; then
@@ -114,41 +144,80 @@ if [ "$REPLY" == "y" ]; then
   open ./Flat.terminal
 fi
 
-read -p "${ARROW} Install nvm (Node Version Manager)? [y/n]: "
+#----------------------------
+# Node Version Manager
+#----------------------------
+if $IS_NVM_INSTALLED; then
+  echo "${ARROW} Node Version Manager already installed!"
+else
+  read -p "${ARROW} Install nvm (Node Version Manager)? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing Node Version Manager..."
-  curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.11/install.sh | bash
-  source ~/.bash_profile
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing Node Version Manager..."
+    curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.11/install.sh | bash
+    source ~/.bash_profile
+
+    IS_NVM_INSTALLED=true
+  fi
 fi
 
-read -p "${ARROW} Install Node.js v9.10.0? [y/n]: "
+#----------------------------
+# Node.js
+#----------------------------
+if $IS_NODE_INSTALLED; then
+  echo "${ARROW} Node.js already installed!"
+else
+  if [ $IS_NVM_INSTALLED == true ]; then
+    read -p "${ARROW} Install Node.js v9.10.0? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing Node.js v9.10.0..."
-  nvm install 9.10.0
+    if [ "$REPLY" == "y" ]; then
+      echo "${ARROW} Installing Node.js v9.10.0..."
+      nvm install 9.10.0
+
+      IS_NODE_INSTALLED=true
+    fi
+  fi
 fi
 
-read -p "${ARROW} Install bundle of npm packages? [y/n]: "
+#----------------------------
+# Npm packages
+#----------------------------
+if $IS_NODE_INSTALLED; then
+  read -p "${ARROW} Install bundle of npm packages? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing npm packages.."
-  install_npm_packages
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing npm packages.."
+    install_npm_packages
+  fi
 fi
 
-read -p "${ARROW} Install bundle of Visual Studio Code extensions? [y/n]: "
+#----------------------------
+# VSCode extensions
+#----------------------------
+if $IS_VSCODE_INSTALLED; then
+  read -p "${ARROW} Install bundle of Visual Studio Code extensions? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Installing Visual Studio Code extensions..."
-  code --install-extension CoenraadS.bracket-pair-colorizer --install-extension PKief.material-icon-theme --install-extension alefragnani.project-manager --install-extension christian-kohler.path-intellisense --install-extension dbaeumer.vscode-eslint --install-extension formulahendry.auto-rename-tag --install-extension mrmlnc.vscode-scss --install-extension msjsdiag.debugger-for-chrome --install-extension techer.open-in-browser --install-extension wayou.vscode-todo-highlight --install-extension xabikos.ReactSnippets
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Installing Visual Studio Code extensions..."
+    code --install-extension CoenraadS.bracket-pair-colorizer --install-extension PKief.material-icon-theme --install-extension alefragnani.project-manager --install-extension christian-kohler.path-intellisense --install-extension dbaeumer.vscode-eslint --install-extension formulahendry.auto-rename-tag --install-extension mrmlnc.vscode-scss --install-extension msjsdiag.debugger-for-chrome --install-extension techer.open-in-browser --install-extension wayou.vscode-todo-highlight --install-extension xabikos.ReactSnippets
+  fi
 fi
 
-read -p "${ARROW} Configure Visual Studio Code settings? [y/n]: "
+#----------------------------
+# VSCode Settings
+#----------------------------
+if $IS_VSCODE_INSTALLED; then
+  read -p "${ARROW} Configure Visual Studio Code settings? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  echo "${ARROW} Configuring Visual Studio Code..."
-  cp settings.json /Users/${USER}/Library/Application\ Support/Code/User
+  if [ "$REPLY" == "y" ]; then
+    echo "${ARROW} Configuring Visual Studio Code..."
+    cp settings.json /Users/${USER}/Library/Application\ Support/Code/User
+  fi
 fi
+
+#----------------------------
+# .gitconfig
+#----------------------------
 
 read -p "${ARROW} Configure .gitconfig? [y/n]: "
 
@@ -157,6 +226,10 @@ if [ "$REPLY" == "y" ]; then
   cp .gitconfig ~
 fi
 
+#----------------------------
+# Spectacle shortcuts
+#----------------------------
+
 read -p "${ARROW} Configure Spectacle shotrcuts? [y/n]: "
 
 if [ "$REPLY" == "y" ]; then
@@ -164,15 +237,35 @@ if [ "$REPLY" == "y" ]; then
   cp -r spectacle.json ~/Library/Application\ Support/Spectacle/Shortcuts.json 2> /dev/null
 fi
 
-read -p "${ARROW} Set up firmware password? [y/n]: "
+#----------------------------
+# Firmware password
+#----------------------------
+if [[ $(sudo firmwarepasswd -check) =~ "Password Enabled: Yes" ]]; then
+  echo "${ARROW} Firmware password is already set up!"
+else 
+  read -p "${ARROW} Set up firmware password? [y/n]: "
 
-if [ "$REPLY" == "y" ]; then
-  if [[ $(sudo firmwarepasswd -check) =~ "Password Enabled: Yes" ]]; then
-    echo "Firmware password is already set up!"
-  else
+  if [ "$REPLY" == "y" ]; then
     sudo firmwarepasswd -setpasswd -setmode command
   fi
 fi
+
+#----------------------------
+# Computer name
+#----------------------------
+read -p "${ARROW} Set computer name? [y/n]: "
+
+if [ "$REPLY" == "y" ]; then
+  read -p 'Please enter computer name: ' uservar
+
+  sudo scutil --set ComputerName $uservar
+  sudo scutil --set HostName $uservar
+  sudo scutil --set LocalHostName $uservar
+fi
+
+#----------------------------
+# macOS Defaults
+#----------------------------
 
 read -p "${ARROW} Configure macOS Defaults? [y/n]: "
